requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
idna==2.8
Pillow==7.0.0
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
pyuploadcare==2.6.0
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
